+++ src/libs/relay/conduit_relay_io_blueprint.cpp
@@ -733,9 +733,9 @@ void write_mesh(const Node &mesh,
     {
         opts_file_style = opts["file_style"].as_string();
 
-        if(opts_suffix != "default" && 
-           opts_suffix != "root_only" &&
-           opts_suffix != "multi_file" )
+        if(opts_file_style != "default" && 
+           opts_file_style != "root_only" &&
+           opts_file_style != "multi_file" )
         {
             CONDUIT_ERROR("write_mesh invalid file_style option: \"" 
                           << opts_file_style << "\"\n"
@@ -843,7 +843,7 @@ void write_mesh(const Node &mesh,
             {
                 static std::map<std::string,int> counters;
                 CONDUIT_INFO("Blueprint save: no 'state/cycle' present."
-                            " Defaulting to counter");
+                             " Defaulting to counter");
                 cycle = counters[path];
                 counters[path]++;
             }
@@ -852,7 +852,7 @@ void write_mesh(const Node &mesh,
                 opts_suffix = "none";
             }
         }
-        else
+        else if(opts_suffix == "default")
         {
             cycle = dom["state/cycle"].to_int();
             opts_suffix = "cycle";
@@ -1754,126 +1754,6 @@ void read_mesh(const std::string &root_file_path,
 //-----------------------------------------------------------------------------
 // -- end conduit::relay::io --
 //-----------------------------------------------------------------------------
-
-
-//-----------------------------------------------------------------------------
-// -- begin conduit::relay::<mpi>::io_blueprint -- (DEPRECATED)
-//-----------------------------------------------------------------------------
-namespace io_blueprint
-{
-
-//---------------------------------------------------------------------------//
-// DEPRECATED
-//---------------------------------------------------------------------------//
-void
-save(const Node &mesh,
-     const std::string &path
-     CONDUIT_RELAY_COMMUNICATOR_ARG(MPI_Comm comm))
-{
-#ifdef CONDUIT_RELAY_IO_MPI_ENABLED
-    save(mesh,
-         path,
-         relay::mpi::io::blueprint::detail::identify_protocol(path),
-         comm);
-#else
-    save(mesh,
-         path,
-         relay::io::blueprint::detail::identify_protocol(path));
-#endif
-}
-
-//---------------------------------------------------------------------------//
-// DEPRECATED
-//---------------------------------------------------------------------------//
-void
-save(const Node &mesh,
-     const std::string &path,
-     const std::string &protocol
-     CONDUIT_RELAY_COMMUNICATOR_ARG(MPI_Comm /*comm*/))
-{
-    // TODO: Add support for yaml protocol
-    Node info;
-    if(protocol != "json" && protocol != "hdf5")
-    {
-        CONDUIT_ERROR("Blueprint I/O doesn't support '" << protocol << "' outputs; "
-                      "output type must be 'blueprint_root' (JSON) or 'blueprint_root_hdf5' (HDF5): " <<
-                      "Failed to save mesh to path " << path);
-    }
-
-    if(!::conduit::blueprint::mesh::verify(mesh, info))
-    {
-        CONDUIT_ERROR("Given node isn't a valid Blueprint mesh: " <<
-                      "Failed to save mesh to path " << path);
-    }
-
-    // NOTE(JRC): The code below is used in lieu of `blueprint::mesh::to_multi_domain`
-    // because the official Blueprint function produces results that are incompatible
-    // with HDF5 outputs (because they include Conduit lists instead of dictionaries).
-    Node index;
-    if(::conduit::blueprint::mesh::is_multi_domain(mesh))
-    {
-        index["data"].set_external(mesh);
-    }
-    else
-    {
-        index["data/mesh"].set_external(mesh);
-    }
-
-    Node &bpindex = index["blueprint_index"];
-    {
-        NodeConstIterator domain_iter = index["data"].children();
-        while(domain_iter.has_next())
-        {
-            const Node &domain = domain_iter.next();
-            const std::string domain_name = domain_iter.name();
-
-            // NOTE: Skip all domains containing one or more mixed-shape topologies
-            // because this type of mesh isn't fully supported yet.
-            bool is_domain_index_valid = true;
-            NodeConstIterator topo_iter = domain["topologies"].children();
-            while(topo_iter.has_next())
-            {
-                const Node &topo = topo_iter.next();
-                is_domain_index_valid &= (
-                    !::conduit::blueprint::mesh::topology::unstructured::verify(topo, info) ||
-                    !topo["elements"].has_child("element_types"));
-            }
-
-            if(is_domain_index_valid)
-            {
-                ::conduit::blueprint::mesh::generate_index(
-                    domain,domain_name,1,bpindex[domain_name]);
-            }
-        }
-    }
-
-    if(bpindex.number_of_children() == 0)
-    {
-        CONDUIT_INFO("No valid domains in given Blueprint mesh: " <<
-                     "Skipping save of mesh to path " << path);
-    }
-    else
-    {
-        index["protocol/name"].set(protocol);
-        index["protocol/version"].set(CONDUIT_VERSION);
-
-        index["number_of_files"].set(1);
-        index["number_of_trees"].set(1);
-        index["file_pattern"].set(path);
-        index["tree_pattern"].set((protocol == "hdf5") ? "data/" : "data");
-
-        relay::io::save(index,path,protocol);
-    }
-}
-
-
-
-//-----------------------------------------------------------------------------
-}
-//-----------------------------------------------------------------------------
-// -- end conduit::relay::<mpi>::io_blueprint --
-//-----------------------------------------------------------------------------
-
 
 
 #ifdef CONDUIT_RELAY_IO_MPI_ENABLED